fix: rename the _bentoml_sdk.api module
